+++ hermod-tracing-core/test/Hermod/Tracing/Test/Unit/Documentation.hs
@@ -40,7 +40,7 @@ config1 = TraceConfig {
       , tofVerbosity = Minimum
       , tofMaxReconnectDelay = 60
       }
-    , tcNodeName = Nothing
+    , tcApplicationName = Nothing
     , tcMetricsPrefix = Just "cardano"
     , tcPeriodicTracers = Map.empty
     , tcPrometheusSimpleRun = Nothing